+++ xfinity-usage/xfinity_logger.py
@@ -1,7 +1,12 @@
 import base64
 import colorlog
 import logging
-from xfinity_globals import *
+import os
+
+
+_LOG_LEVEL = os.environ.get('LOG_LEVEL', 'INFO').upper()
+_DEBUG_LOGGER_FILE = '/config/xfinity.log'
+
 
 color_log_handler = colorlog.StreamHandler()
 color_log_handler.setFormatter(colorlog.ColoredFormatter(
@@ -19,11 +24,11 @@ color_log_handler.setFormatter(colorlog.ColoredFormatter(
 # logger = logging.getLogger(__name__)
 logger = colorlog.getLogger(__name__)
 logger.addHandler(color_log_handler)
-logger.setLevel(LOG_LEVEL)
+logger.setLevel(_LOG_LEVEL)
 debug_formatter = logging.Formatter(fmt='%(asctime)s.%(msecs)03d %(levelname)s: %(message)s', datefmt='%Y-%m-%dT%H:%M:%S')
 
-if LOG_LEVEL == 'DEBUG':
-    file_handler = logging.FileHandler(DEBUG_LOGGER_FILE,mode='w')
+if _LOG_LEVEL == 'DEBUG':
+    file_handler = logging.FileHandler(_DEBUG_LOGGER_FILE,mode='w')
     file_handler.setFormatter(debug_formatter)
     logger.addHandler(file_handler)