+++ xfinity-usage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.5.1
+
+- Obfuscate data in log output
+- If plan data is not returned, download/upload speeds will be set to -1 [#70](https://github.com/thor0215/hassio-xfinity-usage/issues/70)
+- Updated script default POLLING_RATE to 3600 seconds instead 300
+- Added tenacity retry if there is a ConnectionError
+
 ## 0.1.5
 
 - Now pulling data from SelfHelp API, fall back to GraphQL